Improve rules
We're doing package builds in a pbuilder environment that gets
created with a clean tree, so there is no need to run the
`clean` target. That should eliminate one of the two calls to
remakefw. Also we don't need to run tests during a package build
(which we probably didn't do, but this clarifies it).

See also http://build-common.alioth.debian.org/cdbs-doc.html.

Change-Id: I0db03f1ca5fa5991d7a8aa9dc91aa491e95a8f67